access control implemented across all routes

+++ routes/user-deactivated-successfully.js
@@ -19,7 +19,10 @@ router.get("/", async (req, res) => {
     res.redirect("/signout");
   } else {
     utils.setSecurityHeaders(res, beis_url_accessmanagement);
-
+    if (!["BEIS Administrator", "Granting Authority Administrator"].includes(ssn.dashboard_roles)) {
+      return res.render("bulkupload/notAuthorized");
+    }
+    
     try {
       const deleteUser = await axios.delete(
         beis_url_accessmanagement +
